pause & gui reset

+++ RvA/src/states/LostState.cpp
@@ -9,7 +9,7 @@ LostState::LostState(Game& game)
 
 void LostState::draw(Game& game)
 {
-	DrawText("YOU LOST", game.getTexSize().x / 2 - float(MeasureText("YOU LOST!", 20)) / 2, game.getTexSize().y / 2 - 100, 20, WHITE);
+	DrawText("YOU LOST", int(game.getTexSize().x / 2 - MeasureText("YOU LOST!", 20) / 2), int(game.getTexSize().y / 2 - 100), 20, WHITE);
 	if (GuiButton({ game.getTexSize().x / 2 - btnSize.x / 2, game.getTexSize().y / 2, btnSize.x, btnSize.y }, "Menu"))
 	{
 		game.setState(std::make_unique<MenuState>(game));
